appengine/requirements.txt
dash
plotly
pandas
numpy
Flask
gunicorn
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
folium
db-dtypes
holidays